+++ src/main/java/com/pickcoverage/service/ComputatorService.java
@@ -1,5 +1,7 @@
 package com.pickcoverage.service;
 
+import java.math.BigDecimal;
+
 /**
  * Created by stefanbaychev on 3/30/17.
  */
@@ -12,6 +14,6 @@ public interface ComputatorService {
      * @param riskValueInNumPerc the risk value in num perc
      * @return the double
      */
-    Double computeBasicRiskPremiumValue(Double coverValue, Double riskValueInNumPerc);
+    BigDecimal computeBasicRiskPremiumValue(Double coverValue, Double riskValueInNumPerc);
 
 }
